Rename application.
